make ToCSlice public and add bench demo
