+++ flashfunk-level/Cargo.toml
@@ -37,3 +37,7 @@ cc = "1.0"
 [dev-dependencies]
 criterion = "0.3"
 
+[[bench]]
+name = "tick_generation"
+path = "./benchmark/tick_generation.rs"
+harness = false
